rust: sleep first test working
The problem was related to program header of the ELF file aligment. The
address of it need to be page aligned for xv6 to work. The LLVM linker
seems to not work like GCC even when using max-page-size. For now, the
solution was to do the aligment on the linker script. After that the
rust program worked.